thrid question solved

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -1,8 +1,10 @@
 package ImageHoster.controller;
 
+import ImageHoster.model.Comment;
 import ImageHoster.model.Image;
 import ImageHoster.model.Tag;
 import ImageHoster.model.User;
+import ImageHoster.service.CommentService;
 import ImageHoster.service.ImageService;
 import ImageHoster.service.TagService;
 
@@ -32,6 +34,9 @@ public class ImageController<TexturePaintContext> {
     @Autowired
     private TagService tagService;
 
+    @Autowired
+    private CommentService commentService;
+
 
 
     //This method displays all the images in the user home page after successful login
@@ -43,7 +48,7 @@ public class ImageController<TexturePaintContext> {
     }
 
     //This method is called when the details of the specific image with corresponding title are to be displayed
-    //The logic is to get the image from the databse with corresponding title. After getting the image from the database the details are shown
+    //The logic is to get the image from the database with corresponding title. After getting the image from the database the details are shown
     //First receive the dynamic parameter in the incoming request URL in a string variable 'title' and also the Model type object
     //Call the getImageByTitle() method in the business logic to fetch all the details of that image
     //Add the image in the Model type object with 'image' as the key
@@ -57,6 +62,8 @@ public class ImageController<TexturePaintContext> {
         Image image = imageService.getImageById(imageId);
         model.addAttribute("image", image);
         model.addAttribute("tags", image.getTags());
+        List<Comment> com = commentService.getComment(imageId);
+        model.addAttribute("comments",com);
         return "images/image";
     }
 
@@ -145,9 +152,22 @@ public class ImageController<TexturePaintContext> {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        Image image = imageService.getImage(imageId);
+        User curr_User = (User) session.getAttribute("loggeduser");
+        User imageUser = new User(); //
+        imageUser=image.getUser();
+
+        if(imageUser.getUsername().equals(curr_User.getUsername())) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else{
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("image", image);
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        }
     }
